Break socket code into its own file.
