+++ lib/index.js
@@ -1,0 +1,3 @@
+module.exports = function() {
+	this.Socket = require('./socket.js');
+}
