+++ server.js
@@ -209,13 +209,6 @@ function ping_player(socket){
 		});
 }
 
-function send_score(socket){
-	var elapsed_time = now() - socket.start_time,
-		score = elapsed_time / 60.0;
-
-	socket.emit('score_minutes', score.toFixed(3));
-}
-
 function send_team_online(socket){
 	if (!socket || !socket.team){
 		return;
@@ -244,79 +237,6 @@ function set_update_player_interval(socket){
 	}, 5000);
 }
 
-// connect new socket
-function connect(socket){
-	debug_message('Connecting socket');
-	socket.start_time = now();
-	socket.initials = 'unk';
-	// reconnect(socket); // will create player if it doesn't work.
-	create_player(socket);
-	set_update_player_interval(socket);
-
-	socket.on('set_initials', function(data){
-		if (data && data != socket.initials){
-			socket.initials = data.trim().toUpperCase().substring(0, 3);
-			
-			debug_message('Incoming initials: '+socket.initials);
-			update_initials(socket);
-			reset_score(socket);
-		}
-	});
-
-	socket.on('set_team', function(data){
-		if (socket.team){
-			socket.leave(socket.team);
-		}
-
-		if (data){
-			if (data != socket.team){
-				socket.team = data.trim().toUpperCase().substring(0, 3);
-				socket.join(socket.team);
-				
-				debug_message('Incoming team: '+socket.team);
-				update_team(socket);
-			}
-		} else {
-			socket.team = null;
-			
-			debug_message('Incoming team: null');
-			update_team(socket);
-		}
-	});
-
-	socket.on('message', function(data){
-		if (socket.last_message && (now() - socket.last_message) < 10){
-			return;
-		}
-
-		var message = {
-			initials: socket.initials,
-			message: data
-		};
-
-		if (socket.team){
-			message.team = socket.team;
-			socket.broadcast.to(socket.team).emit('message', message);
-		} else {
-			socket.broadcast.emit('message', message);
-		}
-
-		socket.last_message = now();
-	});
-
-	socket.on('disconnect', function(){
-		disconnect(socket);
-	});
-}
-
-function disconnect(socket){
-	if (socket.update_player_interval){
-		clearInterval(socket.update_player_interval);
-	}
-	debug_message('Player '+socket.initials+' disconnecting.');
-
-	socket.disconnect();
-}
 
 // GO
 // Begin
@@ -328,6 +248,7 @@ var express = require('express'),
 	socket_io = require('socket.io'),
 	mysql = require('mysql'), // https://github.com/felixge/node-mysql/
 	_prompt = require('prompt'); // https://github.com/flatiron/prompt
+require('./lib')();
 debug_message('Dependencies loaded...');
 
 // get password and connection info
@@ -370,7 +291,7 @@ _prompt.get({properties: {
 		io = require('socket.io').listen(server);
 	debug_message('Servers ready...');
 
-	io.sockets.on('connection', connect);
+	io.sockets.on('connection', Socket.connect);
 
 	setInterval(function(){
 		emit_online_players();
